Исправление имени свойства
Перенос интерфейса репозиториев, перенос спецификаций
Валидация наличия оператора
