+++ Source/Applications/Backend/WebAPI/DriverAPI/Controllers/V5/PushNotificationsController.cs
@@ -14,8 +14,8 @@ using System.Linq;
 using System.Net.Mime;
 using System.Threading.Tasks;
 using Vodovoz.Application.FirebaseCloudMessaging;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Cash;
-using Vodovoz.EntityRepositories;
 
 namespace DriverAPI.Controllers.V5
 {
