+++ Source/Applications/Backend/WebAPI/TaxcomEdoApi/TaxcomEdoApiExtensions.cs
@@ -10,6 +10,7 @@ using TaxcomEdoApi.Config;
 using TaxcomEdoApi.Converters;
 using TaxcomEdoApi.Factories;
 using TaxcomEdoApi.Services;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.EntityRepositories;
 using Vodovoz.EntityRepositories.Counterparties;
 using Vodovoz.EntityRepositories.Orders;
