+++ Source/Applications/Desktop/Vodovoz/Filters/ViewModels/CounterpartyJournalFilterViewModel.cs
@@ -7,10 +7,10 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data.Bindings.Collections.Generic;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Client;
 using Vodovoz.Domain.Client.ClientClassification;
 using Vodovoz.Domain.Retail;
-using Vodovoz.EntityRepositories;
 using Vodovoz.ViewModels.Counterparties;
 using Vodovoz.ViewModels.Widgets.Search;
