+++ Source/Applications/Desktop/VodovozViewModels/Factories/IncludeExcludeSalesFilterFactory.cs
@@ -6,6 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Client;
 using Vodovoz.Domain.Client.ClientClassification;
 using Vodovoz.Domain.Employees;
@@ -13,7 +14,6 @@ using Vodovoz.Domain.Goods;
 using Vodovoz.Domain.Orders;
 using Vodovoz.Domain.Organizations;
 using Vodovoz.Domain.Sale;
-using Vodovoz.EntityRepositories;
 using Vodovoz.Extensions;
 using Vodovoz.Presentation.ViewModels.Common;
