+++ Source/Applications/Desktop/VodovozViewModels/Journals/FilterViewModels/Store/WarehouseDocumentsItemsJournalFilterViewModel.cs
@@ -12,6 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Core.Domain.Employees;
 using Vodovoz.Domain.Client;
 using Vodovoz.Domain.Documents;
