+++ Source/Applications/Desktop/VodovozViewModels/Journals/JournalViewModels/Orders/OrdersWithoutShipment/OrderWithoutShipmentForAdvancePaymentViewModel.cs
@@ -17,6 +17,7 @@ using System.Collections.Generic;
 using System.Data.Bindings.Collections.Generic;
 using System.Linq;
 using Vodovoz.Controllers;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Employees;
 using Vodovoz.Domain.Goods;
 using Vodovoz.Domain.Orders;
