+++ Source/Applications/Desktop/VodovozViewModels/Journals/JournalViewModels/Orders/OrdersWithoutShipment/OrderWithoutShipmentForDebtViewModel.cs
@@ -13,6 +13,7 @@ using QS.ViewModels;
 using System;
 using System.Data.Bindings.Collections.Generic;
 using System.Linq;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Orders;
 using Vodovoz.Domain.Orders.Documents;
 using Vodovoz.Domain.Orders.OrdersWithoutShipment;
