+++ Source/Applications/Desktop/VodovozViewModels/Journals/JournalViewModels/Orders/OrdersWithoutShipment/OrderWithoutShipmentForPaymentViewModel.cs
@@ -15,6 +15,7 @@ using QS.ViewModels;
 using System;
 using System.Data.Bindings.Collections.Generic;
 using System.Linq;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Client;
 using Vodovoz.Domain.Goods;
 using Vodovoz.Domain.Orders;
