+++ Source/Applications/Desktop/VodovozViewModels/QualityControl/Reports/NumberOfComplaintsAgainstDriversReportViewModel.cs
@@ -10,9 +10,9 @@ using QS.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Complaints;
 using Vodovoz.Domain.Sale;
-using Vodovoz.EntityRepositories;
 using Vodovoz.Presentation.ViewModels.Common;
 using Vodovoz.Settings.Complaints;
 using Vodovoz.Tools;
