+++ Source/Libraries/Core/Backend/DriverAPI.Library/V5/Services/EmployeeService.cs
@@ -3,8 +3,8 @@ using QS.DomainModel.UoW;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Vodovoz.Core.Domain.Common;
 using Vodovoz.Domain.Employees;
-using Vodovoz.EntityRepositories;
 using Vodovoz.EntityRepositories.Employees;
 
 namespace DriverAPI.Library.V5.Services
