+++ Source/Libraries/Core/Business/Vodovoz.Core.Domain/Pacs/Operator.cs
@@ -1,9 +1,12 @@
-﻿using QS.DomainModel.Entity;
+﻿using Microsoft.Extensions.DependencyInjection;
+using QS.DomainModel.Entity;
 using QS.DomainModel.Entity.EntityPermissions;
+using QS.DomainModel.UoW;
 using QS.HistoryLog;
-using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using Vodovoz.Core.Domain.Common;
 
 namespace Vodovoz.Core.Domain.Pacs
 {
@@ -39,9 +42,17 @@ namespace Vodovoz.Core.Domain.Pacs
 			set => SetField(ref _pacsEnabled, value);
 		}
 
-
 		public virtual IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
 		{
+			var unitOfWorkFactory = validationContext.GetService<IUnitOfWorkFactory>();
+
+			var operatorRepository = validationContext.GetService<IGenericRepository<Operator>>();
+
+			if(operatorRepository.GetValue(unitOfWorkFactory.CreateWithoutRoot("Валидация"), o => o.Id, o => o.Id == Id).Any())
+			{
+				yield return new ValidationResult("Оператор уже существует", new[] { nameof(Id) });
+			}
+
 			if(Id == 0)
 			{
 				yield return new ValidationResult("Необходимо выбрать оператора", new[] { nameof(Operator) });
@@ -50,53 +61,6 @@ namespace Vodovoz.Core.Domain.Pacs
 			if(WorkShift == null)
 			{
 				yield return new ValidationResult("Необходимо выбрать смену", new[] { nameof(WorkShift) });
-			}
-		}
-	}
-
-	[Appellative(Gender = GrammaticalGender.Feminine,
-		NominativePlural = "смены",
-		Nominative = "смена")]
-	[EntityPermission]
-	[HistoryTrace]
-	public class WorkShift : PropertyChangedBase, IDomainObject, IValidatableObject
-	{
-		private int _id;
-		private string _name;
-		private TimeSpan _duration;
-
-		[Display(Name = "Код")]
-		public virtual int Id
-		{
-			get => _id;
-			set => SetField(ref _id, value);
-		}
-
-		[Display(Name = "Название")]
-		public virtual string Name
-		{
-			get => _name;
-			set => SetField(ref _name, value);
-		}
-
-
-		[Display(Name = "Длительность")]
-		public virtual TimeSpan Duration
-		{
-			get => _duration;
-			set => SetField(ref _duration, value);
-		}
-
-		public virtual IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
-		{
-			if(Duration <= TimeSpan.Zero)
-			{
-				yield return new ValidationResult("Длительность должна быть больше нуля", new[] { nameof(Duration) });
-			}
-
-			if(string.IsNullOrWhiteSpace(Name))
-			{
-				yield return new ValidationResult("Название должно быть заполнено", new[] { nameof(Name) });
 			}
 		}
 	}
