+++ Source/Libraries/Core/Business/VodovozBusiness/EntityRepositories/GenericRepository.cs
@@ -4,7 +4,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using Vodovoz.Specifications;
+using Vodovoz.Core.Domain.Common;
+using Vodovoz.Core.Domain.Specifications;
 
 namespace Vodovoz.EntityRepositories
 {
