+++ Source/Libraries/Core/Business/VodovozBusiness/Specifications/Orders/EdoContainers/EdoContainerSpecification.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq.Expressions;
+using Vodovoz.Core.Domain.Specifications;
 using Vodovoz.Domain.Orders.Documents;
 
 namespace Vodovoz.Specifications.Orders.EdoContainers
