+++ Source/Libraries/Core/Business/VodovozBusiness/VodovozBusiness.csproj
@@ -605,7 +605,6 @@
     <Compile Include="EntityRepositories\Goods\NomenclaturePricesRepository.cs" />
     <Compile Include="EntityRepositories\HistoryChanges\IArchivedHistoryChangesRepository.cs" />
     <Compile Include="EntityRepositories\HistoryChanges\ArchivedHistoryChangesRepository.cs" />
-    <Compile Include="EntityRepositories\IGenericRepository.cs" />
     <Compile Include="EntityRepositories\Logistic\CachedDistanceRepository.cs" />
     <Compile Include="EntityRepositories\Logistic\CarEventRepository.cs" />
     <Compile Include="EntityRepositories\Logistic\CarRepository.cs" />
@@ -1000,12 +999,6 @@
     <Compile Include="Services\Logistics\IRouteListService.cs" />
     <Compile Include="Services\Logistics\RouteListAcceptStatus.cs" />
     <Compile Include="Services\RevenueServiceInfo.cs" />
-    <Compile Include="Specifications\DynamicExpressionSpecification.cs" />
-    <Compile Include="Specifications\ExpressionSpecification.cs" />
-    <Compile Include="Specifications\ExpressionSpecificationAndExtension.cs" />
-    <Compile Include="Specifications\ExpressionSpecificationNotExtension.cs" />
-    <Compile Include="Specifications\ExpressionSpecificationOrExtension.cs" />
-    <Compile Include="Specifications\ISpecification.cs" />
     <Compile Include="Specifications\Orders\EdoContainers\EdoContainerSpecification.cs" />
     <Compile Include="Tools\Comparers\StringOrNullAfterComparer.cs" />
     <Compile Include="Tools\Exceptions\DeliveryPointDistrictNotFoundException.cs" />
